+++ example/test/widget_test.dart
@@ -20,7 +20,7 @@ void main() {
     expect(find.text('1'), findsNothing);
 
     // Tap the '+' icon and trigger a frame.
-    await tester.tap(find.byIcon(FeatherIcons.plus));
+    await tester.tap(find.byIcon(LucideIcons.plus));
     await tester.pump();
 
     // Verify that our counter has incremented.